+++ lib/flutter_fc.dart
@@ -1,1 +1,13 @@
-export 'src/fc.dart';
+export './src/fc_lib.dart';
+export './src/fc_foundation.dart';
+
+export './src/fc_core.dart'
+    show
+        FCWidget,
+        useSetState,
+        useIsMounted,
+        useDidChangeDependencies,
+        useReassemble,
+        useDispose,
+        useDiff,
+        useMemo;