+++ test/compose_test.dart
@@ -1,10 +1,13 @@
+// ignore_for_file: non_constant_identifier_names
+
 import 'package:flutter/material.dart';
-import 'package:flutter_fc/src/fc.dart';
+import 'package:flutter_fc/flutter_fc.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:flutter_fc/src/fc_compat_v0.dart';
 
 void main() {
   testWidgets("setState in useEffect", (tester) async {
-    final Demo = defineFC((props) {
+    final Demo = defineFC((context, props) {
       final (counter, setCounter) = useState(0);
       useEffect(() {
         setCounter(1);
@@ -19,7 +22,7 @@ void main() {
   });
 
   testWidgets("setState in useImperativeHandle", (tester) async {
-    final Child = defineFC((FCRef? ref) {
+    final Child = defineFC((context, Ref? ref) {
       assert(ref != null);
       final (counter, setCounter) = useState(0);
       useImperativeHandle(ref!, () {
@@ -29,10 +32,10 @@ void main() {
       return Text(counter.toString());
     });
 
-    final Parent = defineFC((props) {
+    final Parent = defineFC((context, props) {
       final ref = useRef(0);
       useEffect(() {
-        expect(ref.current, 0);
+        expect(ref.value, 0);
         return () {};
       });
       return Child(props: ref);